StatusOfData refactor
