+++ services/ui-src/src/measures/2022/shared/CommonQuestions/index.tsx
@@ -6,7 +6,7 @@ export * from "./DataSourceCahps";
 export * from "shared/commonQuestions/AdditionalNotes";
 export * from "./OtherPerformanceMeasure";
 export * from "./Reporting";
-export * from "./StatusOfData";
+export * from "shared/commonQuestions/StatusOfData";
 export * from "./WhyAreYouNotReporting";
 export * from "./OptionalMeasureStrat";
 export * from "shared/commonQuestions/MeasurementSpecification";
