+++ services/ui-src/src/shared/types/index.tsx
@@ -4,3 +4,4 @@ export * from "shared/types/TypeMeasureSpecification";
 export * from "shared/types/TypeDataSource";
 export * from "shared/types/TypeDefinitionOfPopulation";
 export * from "shared/types/TypeDateRange";
+export * from "shared/types/TypeStatusOfData";
